new training new adjustments

+++ src/components/AwarenessEducator/NewTraining/NewTrainingTrainingContent.vue
@@ -143,12 +143,16 @@ export default {
       }))
     },
     handleRemove(index) {
-      AwarenessEducatorService.deleteTrainingFile(
-        this.resourceId,
-        this.formData.contentByLanguage[index].languageId
-      ).then(() => {
-        this.formData.contentByLanguage.splice(index, 1)
-      })
+      if (
+        this?.formData?.contentByLanguage[index]?.file ||
+        this?.formData?.contentByLanguage[index]?.filePreviews?.length
+      ) {
+        AwarenessEducatorService.deleteTrainingFile(
+          this.resourceId,
+          this.formData.contentByLanguage[index].languageId
+        )
+      }
+      this.formData.contentByLanguage.splice(index, 1)
     }
   }
 }
